+++ tilebox-datasets/tilebox/datasets/assets/__init__.py
@@ -1,0 +1,11 @@
+from tilebox.datasets.assets.assets import (
+    Asset,
+    AssetCollection,
+    AssetFieldNames,
+    AssetFields,
+    AssetLocation,
+    Band,
+    MediaType,
+)
+
+__all__ = ["Asset", "AssetCollection", "AssetFieldNames", "AssetFields", "AssetLocation", "Band", "MediaType"]